requirements.txt
# Rasa
rasa==3.6.21
rasa-sdk==3.6.2

# LangGraph / LangChain
langgraph==0.2.54
langchain-core==0.3.0
langchain-openai==0.2.0
langsmith==0.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
